Extract interfact for TransactionStarter
