+++ lock-api/src/test/java/com/palantir/lock/client/TransactionStarterTest.java
@@ -173,7 +173,7 @@ public class TransactionStarterTest {
                                 .result(new DisruptorAutobatcher.DisruptorFuture<>("test"))
                                 .build())
                 .collect(toList());
-        TransactionStarter.consumer(lockLeaseService, lockWatchEventCache).accept(elements);
+        BatchingTransactionStarter.consumer(lockLeaseService, lockWatchEventCache).accept(elements);
         return Futures.getUnchecked(Futures.allAsList(Lists.transform(elements, BatchElement::result)));
     }
